+++ src/main/java/com/bitbank/dto/v1/BankAccountTransactionsDTO.java
@@ -1,0 +1,18 @@
+package com.bitbank.dto.v1;
+
+import java.math.BigDecimal;
+import java.time.LocalDateTime;
+
+import lombok.Data;
+
+@Data
+public class BankAccountTransactionsDTO {
+	
+	private Long id;
+	
+	private BigDecimal amount;
+	
+	private String purpose;
+	
+	private LocalDateTime transactionDate;
+}