+++ CHANGELOG.md
@@ -1,2 +1,16 @@
-# 0.1.0
+# 0.2.0
++ Setup tests
++ Add CI/CD
++ Add Bank Account Transactions entity
++ Add Bank Account Transactions repository
++ Add Bank Account Transactions service
++ Add Bank Account Transactions controller + DTO
++ Add Docker for deploy
++ Add Bank Account Transactions service POST
++ Add Bank Account Transactions controller POST
++ Add Ping service GET
++ Add Ping controller GET
++ Add exclusion in pom.xml for 100% coverage
+
+## 0.1.0
 + First commit